src/requirements.txt
Babel==2.7.0
certifi==2019.3.9
chardet==3.0.4
dataclasses==0.6
defusedxml==0.6.0
Django==2.1.7
django-allauth==0.39.1
django-countries==5.4
django-extra-fields==1.2.4
django-phonenumber-field==3.0.1
django-rest-auth==0.9.5
django-slack==5.13.0
djangorestframework==3.9.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.3.0
djongo==1.2.32
dojango==0.5.8
future==0.17.1
idna==2.8
oauthlib==3.0.1
phonenumbers==8.10.16
Pillow==6.1.0
PyJWT==1.7.1
pymongo==3.7.2
python-twitter==3.5
python3-openid==3.1.0
pytz==2018.9
random-password-generator==2.1.0
requests==2.21.0
requests-oauthlib==1.2.0
six==1.12.0
slackclient==1.3.1
sqlparse==0.2.4
urllib3==1.24.1
websocket-client==0.54.0
